requirements.txt
requests==2.24.0
lxml==4.6.1
beautifulsoup4==4.9.3
Pillow==8.0.1
